+++ src/main/java/br/edu/ifpb/dac/stef/projetojpa2/business/service/PersonService.java
@@ -39,7 +39,7 @@ public class PersonService {
     }
     
 
-    public Person getOne(int Cpf){
+    public Person getOne(Long Cpf){
         return personDAO.findByCpf(Cpf);
     }
 }
